curved crop - query params done, added preset link from home page
